main/requirements.txt
Click==8.0.3
Flask==2.0.2
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
google-cloud-storage
google-cloud-datastore
google-api-python-client
google-cloud-pubsub
google-cloud-runtimeconfig
pyyaml
flask-wtf
cryptography==3.2
python-jose[cryptography]==3.0.1
requests==2.22.0
email-validator==1.0.5
google-auth
google-cloud-logging
pyAesCrypt
marshmallow~=3.14.1
